+++ priv/shaders/learnopengl/1_getting_started/3_4_shaders_exercise_1/vertex_shader.glsl
@@ -1,0 +1,12 @@
+#version 330 core
+layout (location = 0) in vec3 aPos;   // the position variable has attribute position 0
+layout (location = 1) in vec3 aColor; // the color variable has attribute position 1
+
+out vec3 ourColor; // output a color to the fragment shader
+
+void main()
+{
+    // Exercise 1 solution: Make the triangle upside-down by negating the y-coordinate
+    gl_Position = vec4(aPos.x, -aPos.y, aPos.z, 1.0);
+    ourColor = aColor; // set ourColor to the input color we got from the vertex data
+}